feat: Admin 페이지 기능 구현 1. category cascade 처리 2. total값 Context 처리하여 검색할 때마다 테이블 로우 개수를 표시하도록 처리 3. search 조건들 값 변경할 때마다 search 되도록 처리(버튼 누르지 않아도) 4. 방 옵션 수정 안되게 readonly 처리, 새 항목은 수정 되게 함 5. 방 옵션 저장, 삭제 후 검색 조건 옵션 데이터(카테고리) 다시 세팅 6. 테이블 데이터 없을 때 No Data 문구 표시(기존에는 스피너 계속 돌아갔음) 7. 검색 버튼 -> reset 버튼으로 변경(검색 조건 reset, 데이터 refresh)

+++ quiz_duel/src/components/adminPage/lib/optionSave.tsx
@@ -6,20 +6,20 @@ import type { OptionData } from "../types/room-options.types";
 
 // 데이터 유효성 검사
 function validation(optionItems: OptionData[]) {
-  // 중복 키 체크
-  const isDuplicate =
-    optionItems.length !== new Set(optionItems.map((data) => data.id)).size;
-  if (isDuplicate) {
-    toast.error("중복된 키는 저장할 수 없습니다.");
-    return false;
-  }
-
   // 공백 체크
   const isBlank = optionItems.some(
     (item) => item.id === "" || item.value === ""
   );
   if (isBlank) {
     toast.error("공란이 있으면 저장할 수 없습니다.");
+    return false;
+  }
+
+  // 중복 키 체크
+  const isDuplicate =
+    optionItems.length !== new Set(optionItems.map((data) => data.id)).size;
+  if (isDuplicate) {
+    toast.error("중복된 키는 저장할 수 없습니다.");
     return false;
   }
 
@@ -30,16 +30,24 @@ const optionSave = async (
   tableName: string,
   keyColumn: string,
   valueColumn: string,
-  optionItems: OptionData[]
+  optionItems: OptionData[],
+  delOptionItems: string[]
 ) => {
   // 유효성 검사
   if (!validation(optionItems)) return;
 
   try {
-    const { error: delError } = await supabase
-      .from(tableName)
-      .delete()
-      .neq(keyColumn, 0);
+    if (delOptionItems.length > 0) {
+      const { error: delError } = await supabase
+        .from(tableName)
+        .delete()
+        .in(keyColumn, delOptionItems);
+
+      if (delError) {
+        toast.error("저장 실패");
+        return;
+      }
+    }
 
     const { data, error: saveError } = await supabase
       .from(tableName)
@@ -51,7 +59,7 @@ const optionSave = async (
       )
       .select();
 
-    if (delError || saveError) {
+    if (saveError) {
       toast.error("저장 실패");
       return;
     }
@@ -59,7 +67,12 @@ const optionSave = async (
     toast.success("저장 성공");
 
     return data.map((item, index) => {
-      return { key: index, id: item[keyColumn], value: item[valueColumn] };
+      return {
+        key: index,
+        id: item[keyColumn],
+        value: item[valueColumn],
+        editable: false,
+      };
     });
   } catch (error) {
     console.error(error);
